backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
telethon==1.34.0
apscheduler==3.10.4
sqlalchemy==2.0.25
python-dotenv==1.0.1
pydantic==2.6.0
pydantic-settings==2.1.0
